requirements.txt
fastapi==0.104.1
onnxruntime_gpu==1.14
opencv_python_headless==4.6.0.66
Pillow==9.2.0
uvicorn==0.23.2
gunicorn
numpy
python-multipart